+++ books/views/editorials_views.py
@@ -1,13 +1,20 @@
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404, redirect, render
 
 from books.forms import EditorialForm
 from books.models import Editorial
 
 
+@login_required
 def content_editorials(request):
     return render(request, 'content_editorials.html')
 
+@login_required
 def create_editorial(request):
+    if not request.user.is_staff:
+        messages.warning(request, 'no autorizado')
+        return redirect('read_editorials')
     if request.method == 'POST':
         form = EditorialForm(request.POST)
         if form.is_valid():
@@ -21,7 +28,9 @@ def create_editorial(request):
 
     return render(request, 'create_editorial.html', context)
 
+@login_required
 def read_editorials(request):
+    
     queryset = Editorial.objects.all()
     context = {
         'form':queryset
@@ -29,6 +38,7 @@ def read_editorials(request):
     return render(request, 'read_editorials.html', context)
 
 
+@login_required
 def detail_editorial(request, editorial_id):
     editorial = Editorial.objects.get(id=editorial_id)
     context = {
@@ -36,8 +46,11 @@ def detail_editorial(request, editorial_id):
     }
     return render(request, 'detail_editorial.html', context)
 
-
+@login_required
 def edit_editorial(request, editorial_id):
+    if not request.user.is_staff:
+        messages.warning(request, 'no autorizado')
+        return redirect('read_editorials')
     editorial = Editorial.objects.get(id=editorial_id)
     if request.method == 'POST':
         form = EditorialForm(request.POST, instance=editorial)
@@ -52,8 +65,11 @@ def edit_editorial(request, editorial_id):
     }
     return render(request, 'edit_editorial.html', context)
 
-
+@login_required
 def delete_editorial(request, editorial_id):
+    if not request.user.is_staff:
+        messages.warning(request, 'no autorizado')
+        return redirect('read_editorials')
     editorial = get_object_or_404(Editorial,id=editorial_id)
     if request.method == 'POST':
         editorial.delete()
